Added reload whitelist function.
An extra nsq consumer called ReloadConsumer was added.
Any request sent to the consumer will force a reload.
The user of this consumer is the ReloadScripts handler which handles PUT
requests to /scripts.

NOTE: The channel name for the ReloadConsumer should probably be unique
so that whenever a reload request is sent out. All QMD workers will
receive it. If the channel name is the same then at least one gets the
request but maybe not all.