Implement requests from issues #6 and #8.
Modifies the Consumers to no longer connect directly to a single NSQD
node through QueueAddr. API changes have occurred in which Worker now
looks for two config settings called QueueAddresses and LookupAddresses.
Both are lists of address strings for NSQD and NSQLookupd respectively.
This allows a worker to communicate with multiple NSQD and NSQLookupd
nodes.