Removed SIGHUP from graceful shutdown signals.
The inclusion of SIGHUP was causing QMD to shutdown whenever a pipe was
broken. For example when deployed in production and a computer was left
tailing nohup.out. If the person with the computer left for lunch
without exiting the session and the computer went to sleep. This would
send a hangup call to QMD when the pipe broke.